Fix persistence of logging files
In a home assistant AddOn there  is only one persistant volume, mounted
on /data. Therefore create the `logs` directory inside the /data volume,
and create a symlink for it from /opt/tplink/EAPController/logs. Because
of this, the `logs` directory is created at runtime in entrypoint.sh,
and no longer at build time in install.sh. The mounted volume has to
be present.